Having the first argument to main() non-int is an error in GCC 4.3.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@28991 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/magnify/Magnify.cpp
@@ -2062,7 +2062,7 @@ TOSMagnify::ColorAtSelection()
 
 
 int
-main(long argc, char* argv[])
+main(int argc, char* argv[])
 {
 	int32 pixelCount = -1;
